Refactor outputs functions for interfaces and update tests
