+++ lib/interfaces.go
@@ -3,6 +3,7 @@ package lib
 import (
 	"context"
 
+	"github.com/aws/aws-sdk-go-v2/service/cloudformation"
 	"github.com/aws/aws-sdk-go-v2/service/ec2"
 )
 
@@ -16,4 +17,18 @@ type EC2DescribeRouteTablesAPI interface {
 
 type EC2DescribeManagedPrefixListsAPI interface {
 	DescribeManagedPrefixLists(ctx context.Context, params *ec2.DescribeManagedPrefixListsInput, optFns ...func(*ec2.Options)) (*ec2.DescribeManagedPrefixListsOutput, error)
+}
+
+type CFNDescribeStacksAPI interface {
+	DescribeStacks(ctx context.Context, params *cloudformation.DescribeStacksInput, optFns ...func(*cloudformation.Options)) (*cloudformation.DescribeStacksOutput, error)
+}
+
+type CFNListImportsAPI interface {
+	ListImports(ctx context.Context, params *cloudformation.ListImportsInput, optFns ...func(*cloudformation.Options)) (*cloudformation.ListImportsOutput, error)
+}
+
+// CFNExportsAPI combines the CloudFormation operations used for export retrieval
+type CFNExportsAPI interface {
+	CFNDescribeStacksAPI
+	CFNListImportsAPI
 }
